move `getTokenBalance` to eth-defi/helpers
