+++ src/lib/wallet/utils.ts
@@ -1,6 +1,3 @@
-import type { Config, GetBalanceParameters, GetBalanceReturnType } from '@wagmi/core';
-import { getBalance } from '@wagmi/core';
-
 /**
  * Return URL for purchasing native currency for a given chain
  */
@@ -18,22 +15,3 @@ export function buyTokenUrl(chainSlug: string, tokenAddress: Address) {
 	const baseUrl = 'https://app.uniswap.org/#/tokens';
 	return `${baseUrl}/${chainSlug}/${tokenAddress}`;
 }
-
-/**
- * Wrapper around @wagmi getBalance that includes token address
- * in the returned object
- */
-export async function getTokenBalance(
-	config: Config,
-	parameters: GetBalanceParameters & {
-		token: Address;
-	}
-) {
-	const balance = await getBalance(config, parameters);
-	return {
-		...balance,
-		address: parameters.token
-	};
-}
-
-export type GetTokenBalanceReturnType = GetBalanceReturnType & { address: Address };
